Update finance
I've update my work hours. I've counted an hour for the general reading of the code, the tools and the website I've done so far. I've added as well the meeting with Framasoft done earlier this month. Also this afternoon I'm meeting with an investor, I'll send feedback tonight for the details.